Improved errors for invalid connection tooltip (#1303)
