+++ src/renderer/components/outputs/OutputContainer.tsx
@@ -1,6 +1,7 @@
 import { Type } from '@chainner/navi';
 import { Box, Center, HStack, Text, Tooltip, chakra } from '@chakra-ui/react';
 import React, { memo, useCallback, useMemo } from 'react';
+import ReactMarkdown from 'react-markdown';
 import { Connection, Handle, Position, useReactFlow } from 'reactflow';
 import { useContext, useContextSelector } from 'use-context-selector';
 import { OutputId } from '../../../common/common-types';
@@ -38,7 +39,7 @@ const RightHandle = memo(
             hasArrow
             borderRadius={8}
             display={showHandle ? 'none' : 'block'}
-            label={`Unable to connect: ${showHandleReason}`}
+            label={<ReactMarkdown>{`Unable to connect: ${showHandleReason}`}</ReactMarkdown>}
             mt={1}
             opacity={showHandle ? 0 : 1}
             openDelay={500}
